% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceMethods.R
\name{a_ch}
\alias{a_ch}
\title{a_ch}
\usage{
a_ch(...)
}
\arguments{
\item{...}{Arguments to convert.}
}
\value{
Character representation of arguments.
}
\description{
A shortcut to \code{as.character}.
}
