% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{AsSegments}
\alias{AsSegments}
\title{AsSegments}
\usage{
AsSegments(.data, ..., suffix = "_end")
}
\arguments{
\item{.data}{Input table.}

\item{...}{Columns to process.}

\item{suffix}{Suffix to the end value column names.}
}
\value{
Modified table.
}
\description{
AsSegments
}
