% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSTools.R
\name{BSTools.DebugPlot}
\alias{BSTools.DebugPlot}
\title{BSTools.DebugPlot}
\usage{
BSTools.DebugPlot(data)
}
\arguments{
\item{data}{A list of tibbles - one per each simulated chain.}
}
\description{
Plots results of JAGS simulation obtained from \code{BSTools.Run2}.
Result is similar to calling \code{plot} on the direct result of
\code{coda.samples}
}

