% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrames.R
\name{Clamp.data.frame}
\alias{Clamp.data.frame}
\title{Clamp.data.frame}
\usage{
\method{Clamp}{data.frame}(.data, .var, .range)
}
\arguments{
\item{.data}{Input \code{data.frame} or \code{tibble}.}

\item{.var}{Variable to clamp.}

\item{.range}{Clamp range.}
}
\value{
\code{.data} whith clamped within \code{.range} column \code{.var}.
}
\description{
Clamp.data.frame
}
