% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R, R/tools.R
\name{clamp.data.frame}
\alias{clamp.data.frame}
\alias{clamp.tbl_df}
\alias{clamp}
\alias{Clamp}
\alias{clamp.numeric}
\title{clamp}
\usage{
\method{clamp}{data.frame}(...)

\method{clamp}{tbl_df}(...)

clamp(...)

Clamp(...)

\method{clamp}{numeric}(...)
}
\arguments{
\item{...}{Parameter.}
}
\value{
Clamped numerics.
}
\description{
clamp
}
