% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_ClipXY.R
\name{ClipXY}
\alias{ClipXY}
\title{ClipXY}
\usage{
ClipXY(.dt, x, y, xlim, ylim, ..., saveWhich = 1)
}
\arguments{
\item{.dt}{Input table.}

\item{x}{\code{x} column name.}

\item{y}{\code{y} column name.}

\item{xlim}{Limits on \code{x}.}

\item{ylim}{Limits on \code{y}.}

\item{...}{Additional columns to preserve.}

\item{saveWhich}{Which value to assign to the non-interpolated
column. Can be 1 (preceeding) and 2 (succeeding)}
}
\description{
ClipXY
}
