% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\docType{class}
\name{Cluster-class}
\alias{Cluster-class}
\alias{Cluster}
\title{Cluster}
\description{
RefClass representing cluster for parallel execution.
}
\section{Fields}{

\describe{
\item{\code{ClusterDesc}}{A \code{SOCKcluster} underkying object.}

\item{\code{IsRegistered}}{A \code{logical} flag indicating
whether the cluster was registered.}

\item{\code{IsDisposed}}{A \code{logical} flag indicating
whrther cluster resources were disposed.}

\item{\code{ID}}{A 'unique' id of the cluster, hex string.}
}}


