% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{contours_2d}
\alias{contours_2d}
\alias{JointDistributionContours}
\title{contours_2d}
\usage{
contours_2d(x, y, prob, n = 30, lims = c(range(x), range(y)))

JointDistributionContours(...)
}
\arguments{
\item{x}{First variable.}

\item{y}{Second variable.}

\item{prob}{Vector of probabilities in (0, 1).}

\item{n}{Size of the \code{MASS::kde2d} estimator's grid.}

\item{lims}{Explicit lims passed to \code{kde2d}.}

\item{...}{For compatibility}
}
\value{
List of contours, one per each probability.
}
\description{
contours_2d
}
