% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{contours_2d_df}
\alias{contours_2d_df}
\title{contours_2d_df}
\usage{
contours_2d_df(df, x, y, prob, n = 30L, lims = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{x}{First variable.}

\item{y}{Second variable.}

\item{prob}{Vector of probabilities in (0, 1).}

\item{n}{Size of the \code{MASS::kde2d} estimator's grid.}

\item{lims}{Explicit lims passed to \code{kde2d}.}
}
\value{
List of contours, one per each probability.
}
\description{
contours_2d_df
}
