% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converters.R
\name{Dec2Degrees}
\alias{Dec2Degrees}
\title{Dec2Degrees}
\usage{
Dec2Degrees(x, sep = " ", simple = TRUE)
}
\arguments{
\item{x}{Vector of values.}

\item{sep}{Separator between degrees, minutes and seconds.}

\item{simple}{If \code{TRUE}, return vector of converted values.
If \code{FALSE}, returns a \code{tibble} with more detatiled infomration.}
}
\value{
Either a vector of length \code{n} o a tibble \code{n x 5}.
}
\description{
Converts \code{DEC}lination to degrees.
}
