% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{DefaultTheme}
\alias{DefaultTheme}
\title{DefaultTheme}
\usage{
DefaultTheme(ticks = unit(-3.5, "pt"), textSz = 10, titleSz = 15)
}
\arguments{
\item{ticks}{Controls the tick size (and direction).}

\item{textSz}{Text font size by default.}

\item{titleSz}{Title (labels') font size by default.}
}
\value{
A theme object.
}
\description{
Generates default theme used in scientific publications.
}
