% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converters.R
\name{Degrees2Dec}
\alias{Degrees2Dec}
\title{Degrees2Deg}
\usage{
Degrees2Dec(x, sep = " ", dig = 3)
}
\arguments{
\item{x}{Vector of values.}

\item{sep}{Separator between degrees, minutes and seconds.}

\item{dig}{Number of decimal digits in seconds.}
}
\value{
A string vector.
}
\description{
Converts numerical degrees to string representation
of \code{DEC}lination.
}
