% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converters.R
\name{Degrees2Ra}
\alias{Degrees2Ra}
\title{Degrees2Ra}
\usage{
Degrees2Ra(x, sep = " ", dig = 3)
}
\arguments{
\item{x}{Vector of values.}

\item{sep}{Separator between hours, minutes and seconds.}

\item{dig}{Number of decimal digits in seconds.}
}
\value{
A string vector.
}
\description{
Converts numerical degrees to string representation
of \code{R}ight \code{A}scention.
}
