% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ex.R
\name{desc_ex}
\alias{desc_ex}
\title{desc_ex}
\usage{
desc_ex(f)
}
\arguments{
\item{f}{A sorting object to inverse (a collection or a function)}
}
\value{
Modified colelction / function
}
\description{
desc_ex
}
