% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%!=\%}
\alias{\%===\%}
\alias{\%!==\%}
\title{Floating-point equality}
\usage{
e1 \%==\% e2

e1 \%!=\% e2

e1 \%===\% e2

e1 \%!==\% e2
}
\arguments{
\item{e1}{LHS}

\item{e2}{RHS}
}
\value{
Logical vector

Logical vector
}
\description{
Returns \code{logical} of size 1, \code{TRUE} if all elements are equal, otherwise \code{FALSE},

Returns \code{logical} of size 1, \code{TRUE} if any elements are unequal, otherwise \code{FALSE},
}
