% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{Expand}
\alias{Expand}
\title{Expand}
\usage{
Expand(x, factor = 1, direction = c(1, 1))
}
\arguments{
\item{x}{Input interval.}

\item{factor}{How large the expansion is.
1 corresponds to 100\% increase in size.}

\item{direction}{In which way to expand.}
}
\value{
Expanded interval.
}
\description{
Expand
}
