% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{FancyStep}
\alias{FancyStep}
\title{FancyStep}
\usage{
FancyStep(range, n = 6, modifier = c(1, 2.5, 5))
}
\arguments{
\item{range}{The range within which the ticks are placed.}

\item{n}{Approximate number of desired ticks.}

\item{modifier}{Preferred tick placements/}
}
\value{
Returns the size of the step.
}
\description{
FancyStep
}
