% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{filter_range}
\alias{filter_range}
\alias{FilterRange}
\title{fitler_range}
\usage{
filter_range(.data, .var, .range, .strict = FALSE)

FilterRange(.data, .var, .range, .strict = FALSE)
}
\arguments{
\item{.data}{Input table.}

\item{.var}{Column to filter.}

\item{.range}{Limits on the column.}

\item{.strict}{\code{logical}; If \code{TRUE}, boundaries are not included.}
}
\value{
Filtered table.
}
\description{
fitler_range
}
