% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrames.R
\name{FilterRange}
\alias{FilterRange}
\title{FitlerRange}
\usage{
FilterRange(.data, .var, .range)
}
\arguments{
\item{.data}{Input table.}

\item{.var}{Column to filter.}

\item{.range}{Limits on the column.}
}
\value{
FIltered table.
}
\description{
FitlerRange
}
