% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GenerateBreaks}
\alias{GenerateBreaks}
\title{GenerateBreaks}
\usage{
GenerateBreaks(range, largeStep, smallStep)
}
\arguments{
\item{range}{Range of breaks (typically a scale of an axis).}

\item{largeStep}{Step for large breaks (the ones with labels).}

\item{smallStep}{Step for small breaks (only ticks).}
}
\description{
Generates large and small breaks within given range.
}
