% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GenerateLog10Breaks.R
\name{GenerateLog10Breaks}
\alias{GenerateLog10Breaks}
\title{GenerateLog10Breaks}
\usage{
GenerateLog10Breaks(ylim, ticks = c(0.2, 0.5, 1, 2, 5))
}
\arguments{
\item{ylim}{Axis limit.}

\item{ticks}{If the limit spans over several orders,
\code{ticks} are used to generate large breaks.}
}
\value{
log10 breaks.
}
\description{
GenerateLog10Breaks
}
