% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GetPlotLims.R
\name{GetPlotLims}
\alias{GetPlotLims}
\title{GetPlotLims}
\usage{
GetPlotLims(plt)
}
\arguments{
\item{plt}{Input \code{ggplot} plot.}
}
\value{
\code{list} with x and y limits.
}
\description{
GetPlotLims
}
