% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{GGCustomTextAnnotation}
\alias{GGCustomTextAnnotation}
\title{GGCustomTextAnnotation}
\usage{
GGCustomTextAnnotation(labels, x, y, xmin = x, xmax = x, ymin = y,
  ymax = y, vjust = 0, hjust = 0, rot = 0, gp = list(gpar()))
}
\arguments{
\item{labels}{Labels.}

\item{xmin, xmax}{X coordinates of labels.}

\item{ymin, ymax}{Y coordinates of labels.}

\item{vjust}{Vertical justification.}

\item{hjust}{Horizontal justification.}

\item{rot}{Rotation angle.}

\item{gp}{Additional graphical parameters.}
}
\value{
List of \code{annotation_custom} objects.
}
\description{
GGCustomTextAnnotation
}
