% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{GGPlot2Grob}
\alias{GGPlot2Grob}
\title{GGPlot2Grob}
\usage{
GGPlot2Grob(plots, innerMar, outerMar, clip = FALSE)
}
\arguments{
\item{plots}{Plots to convert.}

\item{innerMar}{If present, sets inner margins.}

\item{outerMar}{If present, sets outer margins.}

\item{clip}{If \code{TRUE}, clips everything outside of the plot ares
(affects custom tick labels and so on), otherwise, allows contents outside 
of the plot area.}
}
\value{
A collection of \code{grobs}.
}
\description{
Converts \code{ggplot2} objects into \code{gTables}
and sets up margins.
}
