% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GGPlot2GrobEx}
\alias{GGPlot2GrobEx}
\title{GGPlot2GrobEx}
\usage{
GGPlot2GrobEx(plots, clip = FALSE)
}
\arguments{
\item{plots}{Input plots.}

\item{clip}{Clip flag.}
}
\value{
Colelction of grobs.
}
\description{
GGPlot2GrobEx
}
