% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GGPlotCustomTicks}
\alias{GGPlotCustomTicks}
\title{GGPlotCustomTicls}
\usage{
GGPlotCustomTicks(plt, side, breaks, labels, tckSz, trnsf = identity,
  gp = gpar(), rot = 0, delta = 0)
}
\description{
Creates custom ticks with labels. Requires a finished ggproto object.
Can be piped with %>%.
}

