% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GGPlotCustomTicksEx2.R
\name{GGPlotCustomTicksEx2}
\alias{GGPlotCustomTicksEx2}
\title{GGPlotCustomTicksEx}
\usage{
GGPlotCustomTicksEx2(plt, side, breaks, labels, trnsf = identity,
  tckSz = unit(7.5, "pt"), gp = gpar(), rot = 0,
  offset = unit(0.15, "cm"), just = NULL, tickGp = gpar())
}
\arguments{
\item{plt}{Target plot.}

\item{side}{Indicates axis. Can be either \code{1,2,3,4},
\code{"b", "l", "t", "r"}, or strings containing
\code{"bot", "lef", "top", "rig"} substring.}

\item{breaks}{Breaks in the axis scale.}

\item{labels}{Respective text labels.}

\item{trnsf}{Inverse trnsformation function. Used to handle secondary axes.}

\item{tckSz}{Tick size (in units).}

\item{gp}{A \code{gpar} set of parameters passed to the text methods.
Affects rendering of the labels.}

\item{rot}{Rotation of labels, deg.}

\item{offset}{Tick label offset from the axis.}

\item{just}{Overrides default justification.}

\item{tickGp}{\code{gpar} that controls ticks.}
}
\description{
Creates custom ticks with labels. Requires a finished ggproto object.
Can be piped with %>%.
}
