% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GGPlotGetRange}
\alias{GGPlotGetRange}
\title{GGPlotGetRange}
\usage{
GGPlotGetRange(plt)
}
\arguments{
\item{plt}{Plot to measure.}
}
\value{
A \code{list()} of ranges with names \code{"x", "y", "x2", "y2"}.
Some values (like *2) can be \code{NULL} if no respective axis is present.
}
\description{
Builds provided \code{ggplot2} object and determines
actual ranges of all of its axes.
}
