% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GGPlotLims.R
\name{GGPlotLims}
\alias{GGPlotLims}
\title{GGPlotLims}
\usage{
GGPlotLims(plt)
}
\arguments{
\item{plt}{Input \code{ggplot} plot.}
}
\value{
\code{list} with x and y limits.
}
\description{
GGPlotLims
}
