% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GGPlotPanelLabs}
\alias{GGPlotPanelLabs}
\title{GGPlotPanelLabs}
\usage{
GGPlotPanelLabs(
  p,
  labels = "X",
  x = Inf,
  y = Inf,
  hjust = 1,
  vjust = 1,
  gp = gpar()
)
}
\arguments{
\item{p}{Input plots.}

\item{labels}{Labels to place on each plot, one per plot.}

\item{x}{X cordinates of the labels. Can be +-Inf.}

\item{y}{Y cordinates of the labels. Can be +-Inf.}

\item{hjust}{Horizontal justification. + moves right.}

\item{vjust}{Vertical justification. + moves down.}

\item{gp}{Text parameters}
}
\value{
Modifed plots.
}
\description{
GGPlotPanelLabs
}
