% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_fmt.R
\name{glue_fmt}
\alias{glue_fmt}
\alias{GlueFmt}
\title{glue_fmt}
\usage{
glue_fmt(..., .envir = parent.frame())

GlueFmt(..., .envir = parent.frame())
}
\arguments{
\item{...}{Passed to \code{glue::glue}.}

\item{.envir}{Evaluation environment.}
}
\value{
Format-aware interpoalted string.
}
\description{
glue_fmt
}
