% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_fmt.R
\name{glue_fmt_chr}
\alias{glue_fmt_chr}
\title{glue_fmt_chr}
\usage{
glue_fmt_chr(..., .envir = parent.frame())
}
\arguments{
\item{...}{Passed to \code{glue::glue}.}

\item{.envir}{Evaluation environment.}
}
\value{
Format-aware interpoalted string.
}
\description{
glue_fmt_chr
}
