% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%.\%}
\alias{\%.\%}
\title{Dot-product}
\usage{
x \%.\% y
}
\arguments{
\item{x}{LHS.}

\item{y}{RHS.}
}
\value{
A dot product of two vectors.
}
\description{
A \code{vctrs}-typed replacement to \code{base::`\%*\%`}.
}
