% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%??\%}
\alias{\%??\%}
\title{Null/empty-coalescing operator}
\usage{
x \%??\% y
}
\arguments{
\item{x}{Left side of the operator. To be tested.}

\item{y}{Right side of the operator. Is returned if left side is
null or empty.}
}
\value{
Either x or y.
}
\description{
Improves the \code{rlang::\%||\%} operator by
handling also cases of zero-length objects.
}
