% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{\%is\%}
\alias{\%is\%}
\title{\code{is} interfix operator}
\usage{
object \%is\% class
}
\arguments{
\item{object}{Object to test.}

\item{class}{Target type (supports \code{rlang} quosure).}
}
\value{
\code{logical} \code{TRUE} if
\code{object} is of class \code{class}, \code{FALSE} otherwise.
}
\description{
An interfix version of \link{is} method.
}
