% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{\%vec_in\%}
\alias{\%vec_in\%}
\title{vec_in}
\usage{
needles \%vec_in\% haystack
}
\arguments{
\item{needles}{Needles (what to match).}

\item{haystack}{Haystack (where to match).}
}
\value{
Logical vector of matches.
}
\description{
vec_in
}
