% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GrobMarginSet.R
\name{GrobMarginSet}
\alias{GrobMarginSet}
\title{GrobMarginSet}
\usage{
GrobMarginSet(grob, tLab = NULL, rLab = NULL, bLab = NULL,
  lLab = NULL, tAxis = NULL, rAxis = NULL, bAxis = NULL,
  lAxis = NULL, labsList = NULL, axisList = NULL, labsMar = NULL,
  axisMar = NULL)
}
\arguments{
\item{grob}{Input grob.}

\item{tLab}{Top label margin.}

\item{rLab}{Right label margin.}

\item{bLab}{Bottom label margin.}

\item{lLab}{Left label margin.}

\item{tAxis}{Top axis margin.}

\item{rAxis}{Right axis margin.}

\item{bAxis}{Bottom axis margin.}

\item{lAxis}{Left axis margin.}

\item{labsList}{When not NULL, is used instead of other label parameters.}

\item{axisList}{When not NULL, is used instead of other axis parameters.}

\item{labsMar}{A \code{margin} item used instead of \code{labsList}.}

\item{axisMar}{A \code{margin} item used instead of \code{axisList}.}
}
\value{
Modified grob.
}
\description{
GrobMarginSet
}
