% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GrobPlot}
\alias{GrobPlot}
\title{GrobPlot}
\usage{
GrobPlot(grobs, noNewPageDevList = c("pdf"))
}
\arguments{
\item{grobs}{Figures to plot.}

\item{noNewPageDevList}{Names of devices that produce empty page
if \code{grid::grid.newage()} is called before first call
of \code{grid::grid.draw(...)}.}
}
\description{
Plots provide \code{gTable} objects, one per page
}
