% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GrobsArrange.R
\name{GrobsArrange}
\alias{GrobsArrange}
\title{GrobsArrange}
\usage{
GrobsArrange(grobs, ncol, nrow = length(grobs)\%/\%ncol,
  tLab = unit(0.6, "cm"), rLab = unit(0.6, "cm"), bLab = unit(0.6,
  "cm"), lLab = unit(0.6, "cm"), tAxis = unit(0.6, "cm"),
  rAxis = unit(0.6, "cm"), bAxis = unit(0.6, "cm"), lAxis = unit(0.6,
  "cm"), vGap = unit(0.1, "cm"), hGap = unit(0.1, "cm"),
  labsList = NULL, axisList = NULL, labsMar = NULL, axisMar = NULL,
  topLab = NULL, bottomLab = NULL, leftLab = NULL, rightLab = NULL)
}
\arguments{
\item{grobs}{Input grobs.}

\item{ncol}{Number of columns.}

\item{nrow}{Number of rows.}

\item{tLab}{Top label margin.}

\item{rLab}{Right label margin.}

\item{bLab}{Bottom label margin.}

\item{lLab}{Left label margin.}

\item{tAxis}{Top axis margin.}

\item{rAxis}{Right axis margin.}

\item{bAxis}{Bottom axis margin.}

\item{lAxis}{Left axis margin.}

\item{vGap}{Vertical gap between plots.}

\item{hGap}{Horizontal gap between plots.}

\item{labsList}{Label margins in form of named list.
If not \code{NULL}, used instead of other provided values.}

\item{axisList}{Axis margins in form of named list.
If not \code{NULL}, used instead of other provided values.}

\item{labsMar}{A \code{margin} item used instead of \code{labsList}.}

\item{axisMar}{A \code{margin} item used instead of \code{axisList}.}

\item{topLab}{\code{textGrob} which can act as shared top axis label.}

\item{bottomLab}{\code{textGrob} which can act as shared bottom axis label.}

\item{leftLab}{\code{textGrob} which can act as shared left axis label.}

\item{rightLab}{\code{textGrob} which can act as shared right axis label.}
}
\value{
GTable of arranged grobs.
}
\description{
GrobsArrange
}
