% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GrobSetGaps}
\alias{GrobSetGaps}
\title{GrobSetGaps}
\usage{
GrobSetGaps(
  grob,
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL,
  asList = NULL,
  asMargin = NULL
)
}
\arguments{
\item{grob}{Inout grob.}

\item{top}{Top gap.}

\item{right}{Right gap.}

\item{bottom}{Bottom gap.}

\item{left}{Left gap.}

\item{asList}{When not NULL, used instead of other parameters to set gaps.}

\item{asMargin}{When not NULL, used instead of other parameters to set gaps.}
}
\value{
Grob.
}
\description{
GrobSetGaps
}
