% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{GrobSizeGet}
\alias{GrobSizeGet}
\title{GrobSizeGet}
\usage{
GrobSizeGet(grob, name, id)
}
\arguments{
\item{grob}{Input grob.}

\item{name}{Name of the grob.}

\item{id}{Alternatively, id of the grob.
Useful when there are multiple grobs of the same name.}
}
\value{
Size of the grob.
}
\description{
GrobSizeGet
}
