% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{GrobSizeSet}
\alias{GrobSizeSet}
\title{GrobSizeSet}
\usage{
GrobSizeSet(grob, name, width = NULL, height = NULL, id)
}
\arguments{
\item{grob}{Input grob.}

\item{name}{Name of the grob.}

\item{width}{Width to set.}

\item{height}{Height to set.}

\item{id}{Alternatively, id of the grob.
Useful when there are multiple grobs of the same name.}
}
\value{
Modfied grob.
}
\description{
GrobSizeSet
}
