% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{GrobZero}
\alias{GrobZero}
\title{GrobZero}
\usage{
GrobZero(grob, name)
}
\arguments{
\item{grob}{Grob collection.}

\item{name}{Name of grob to remove.}
}
\value{
Modified grob collection.
}
\description{
GrobZero
}
