% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tools.R
\name{has_size}
\alias{has_size}
\title{has_size}
\usage{
has_size(x, len)
}
\arguments{
\item{x}{Tested object}

\item{len}{Expected size}
}
\value{
Result of the \code{vctrs::vec_size(x) == len} test
}
\description{
has_size
}
\examples{
\dontrun{assertthat::assert_that(has_size(1, 2))}
}
