% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{if_else_weak}
\alias{if_else_weak}
\title{if_else_weak}
\usage{
if_else_weak(condition, true, false)
}
\arguments{
\item{condition}{Condition to test}

\item{true}{Returned if \code{condition} is \code{TRUE}}

\item{false}{Returned otherwise}
}
\description{
if_else_weak
}
