% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceMethods.R
\name{Intersect}
\alias{Intersect}
\title{Intersect}
\usage{
Intersect(x, y, tol = 0.1)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}

\item{tol}{The tolerance level.}
}
\value{
Indices of first and second vector.
These elements are found to be equal within \code{tol}
}
\description{
Intersect
}
