% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tools.R
\name{is_numeric}
\alias{is_numeric}
\title{is_numeric}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{Vector to test}
}
\value{
\code{TRUE} if \code{x} is either \code{double} or \code{integer}.
}
\description{
is_numeric
}
