% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{left_join_safe}
\alias{left_join_safe}
\alias{inner_join_safe}
\alias{right_join_safe}
\alias{full_join_safe}
\alias{nest_join_safe}
\alias{semi_join_safe}
\alias{anti_join_safe}
\title{join}
\usage{
left_join_safe(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"),
  ...)

inner_join_safe(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

right_join_safe(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

full_join_safe(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"),
  ...)

nest_join_safe(x, y, by = NULL, copy = FALSE, keep = FALSE,
  name = NULL, ...)

semi_join_safe(x, y, by = NULL, copy = FALSE, ...)

anti_join_safe(x, y, by = NULL, copy = FALSE, ...)
}
\arguments{
\item{x, y}{tbls to join}

\item{by}{a character vector of variables to join by.}

\item{copy}{Should tables be copied.}

\item{suffix}{Column names suffixes if names are the same.}

\item{...}{Other params}

\item{keep}{If TRUE the by columns are kept in the nesting joins.}

\item{name}{Used in nested join.}
}
\value{
Joined table using respective \code{dplyr::*_join()}
}
\description{
join
}
