% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{join_cnd}
\alias{join_cnd}
\alias{inner_join_cnd}
\alias{left_join_cnd}
\alias{right_join_cnd}
\alias{full_join_cnd}
\title{join_cnd}
\usage{
join_cnd(
  left,
  right,
  ...,
  .type = "inner",
  .selector = "all",
  .suffix = c("__l", "__r"),
  .enforce_suffix = FALSE
)

inner_join_cnd(
  left,
  right,
  ...,
  .selector = "all",
  .suffix = c("__l", "__r"),
  .enforce_suffix = FALSE
)

left_join_cnd(
  left,
  right,
  ...,
  .selector = "all",
  .suffix = c("__l", "__r"),
  .enforce_suffix = FALSE
)

right_join_cnd(
  left,
  right,
  ...,
  .selector = "all",
  .suffix = c("__l", "__r"),
  .enforce_suffix = FALSE
)

full_join_cnd(
  left,
  right,
  ...,
  .selector = "all",
  .suffix = c("__l", "__r"),
  .enforce_suffix = FALSE
)
}
\arguments{
\item{left}{LHS table.}

\item{right}{RHS table}

\item{...}{Joining conditions, where \code{.x} refers to RHS table and
\code{.y} referes to LHS table. E.g. \code{.x$hp > .y$mpg}.
Comma-separated conditions are \code{`&`}.}

\item{.type}{Type of the performed join.}

\item{.selector}{If multiple matches, which RHS to select}

\item{.suffix}{Suffixes that are added to the repeating columns.}

\item{.enforce_suffix}{If \code{TRUE}, enforces suffixes on all columns.}
}
\description{
join_cnd
}
