% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{JointDistribution}
\alias{JointDistribution}
\title{JointDistribution}
\usage{
JointDistribution(x, y, prob, n = 30)
}
\arguments{
\item{x}{First variable.}

\item{y}{Second variable.}

\item{prob}{Vector of probabilities in (0, 1).}

\item{n}{Size of the \code{MASS::kde2d} estimator's grid.}
}
\value{
List of contours, one per each probability.
}
\description{
JointDistribution
}
