% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{JointDistributionContours}
\alias{JointDistributionContours}
\title{JointDistributionContours}
\usage{
JointDistributionContours(x, y, prob, n = 30, lims = c(range(x),
  range(y)))
}
\arguments{
\item{x}{First variable.}

\item{y}{Second variable.}

\item{prob}{Vector of probabilities in (0, 1).}

\item{n}{Size of the \code{MASS::kde2d} estimator's grid.}

\item{lims}{Explicit lims passed to \code{kde2d}.}
}
\value{
List of contours, one per each probability.
}
\description{
JointDistributionContours
}
