% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_LabelsRep.R
\name{labels_rep}
\alias{labels_rep}
\alias{LabelsRep}
\title{labels_rep}
\usage{
labels_rep(labels, norm_range = c(-2, 2), is_tex = FALSE,
  exp_dig_plc = 0, omit_ones = TRUE, every = 1,
  same_digit_count = FALSE)

LabelsRep(...)
}
\arguments{
\item{labels}{Input numeric vector.}

\item{norm_range}{Range within which the numers are
represented without exponential notation (fixed).}

\item{is_tex}{If \code{TRUE}, writes output as TeX compatible strings,
otherwiase, uses \code{expression}.}

\item{exp_dig_plc}{How many digits should be left in the exponential notation.}

\item{omit_ones}{If \code{TRUE}, all powers of 10 are written
without leading 1}

\item{every}{Samples labels. All other labels, except every \code{every} one
are empty strings. Works good when there is no space to put all the labels.}

\item{same_digit_count}{If \code{TRUE}, enforces same amount of
dec. digits across all items.}

\item{...}{Parameter list}
}
\value{
Textual represntation of the numeric vector.
}
\description{
labels_rep
}
