% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_LabelsRep.R
\name{LabelsRep}
\alias{LabelsRep}
\title{LabelsRep}
\usage{
LabelsRep(labels, normRange = c(-2, 2), isTex = FALSE, expDigPlc = 0,
  omitOnes = TRUE, every = 1)
}
\arguments{
\item{labels}{Input numeric vector.}

\item{normRange}{Range within which the numers are
represented without exponential notation (fixed).}

\item{isTex}{If \code{TRUE}, writes output as TeX compatible strings,
otherwiase, uses \code{expression}.}

\item{expDigPlc}{How many digits should be left in the exponential notation.}

\item{omitOnes}{If \code{TRUE}, all powers of 10 are written
without leading 1}

\item{every}{Samples labels. All other labels, except every \code{every} one
are empty strings. Works good when there is no space to put all the labels.}
}
\value{
Textual represntation of the numeric vector.
}
\description{
LabelsRep
}
