% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{Lin}
\alias{Lin}
\title{Lin}
\usage{
Lin(x0, x, y)
}
\arguments{
\item{x0}{Where to interpolate.}

\item{x}{Arguments (size 2).}

\item{y}{Values (size 2).}
}
\value{
Interpolated value between two provided.
}
\description{
Lin
}
