% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{list_collapse_dfr}
\alias{list_collapse_dfr}
\title{List collapse}
\usage{
list_collapse_dfr(data, to = Name, ptype = factor(), modifier = identity)
}
\arguments{
\item{data}{List of \code{data.frame}s.}

\item{to}{Name of the column to where list names will go.}

\item{ptype}{Casts names to the target type. By default, \code{factor()};
when \code{NULL}, perofrms no cast.}

\item{modifier}{An \code{rlang}-style function/lambda that converts names.
Can e.g. be used to parse names into numbers and than cast column to \code{ptype} type.}
}
\value{
A single \code{data.frame}.
}
\description{
Collapses named list of \code{data.frame} rowwise, assigning
list names to the column named \code{to}.
}
