% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{Log10Ceiling}
\alias{Log10Ceiling}
\title{Log10Ceiling}
\usage{
Log10Ceiling(x)
}
\arguments{
\item{x}{Input numeric vector.}
}
\value{
Closest power of 10 that is greater than or equal to the number.
}
\description{
Log10Ceiling
}
