% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{Log10Floor}
\alias{Log10Floor}
\title{Log10Floor}
\usage{
Log10Floor(x)
}
\arguments{
\item{x}{Input numeric vector.}
}
\value{
Closest power of 10 that is smaller than or equal to the number.
}
\description{
Log10Floor
}
