% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa_phase.R
\name{lsa_phase}
\alias{lsa_phase}
\alias{LSAPhase}
\title{lsa_phase}
\usage{
lsa_phase(w, t)

LSAPhase(w, t)
}
\arguments{
\item{w}{(Angular) frequencies.}

\item{t}{Time.}
}
\value{
Phase shifts in form of a vector that is utilized by the
\code{LSAPeriodogram}.
}
\description{
lsa_phase
}
