% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAPeriodogram.R
\name{LSAPeriodogram}
\alias{LSAPeriodogram}
\title{LSAPeriodogram}
\usage{
LSAPeriodogram(w, t, x, tau = NA)
}
\arguments{
\item{w}{(Angular) frequencies.}

\item{t}{Moments of time at which \code{x} were observed.}

\item{x}{Observations}

\item{tau}{If \code{NA}, \code{LSAPhase} is used to calculate phases.
Can be used to caches values for a given set of \code{w} and \code{t}.}
}
\value{
Numeric vector of length equal to length of \code{w}.
}
\description{
LSAPeriodogram
}
