% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAPeriodogramEx.R
\name{LSAPeriodogramEx}
\alias{LSAPeriodogramEx}
\title{LSAPeriodogram}
\usage{
LSAPeriodogramEx(.data, t, x, w, tau = NA)
}
\arguments{
\item{.data}{Input data tibble.}

\item{t}{Time column.}

\item{x}{Measurement column.}

\item{w}{(Angular) frequencies}

\item{tau}{Phase shifts. If \code{NA},
\code{RLibs::LSAPhase} is used to obtain values.}
}
\value{
Table containg parameters of the LS periodogram.
}
\description{
LSAPeriodogram
}
