% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAPhase.R
\name{LSAPhase}
\alias{LSAPhase}
\title{LSAPhase}
\usage{
LSAPhase(w, t)
}
\arguments{
\item{w}{(Angular) frequencies.}

\item{t}{Time.}
}
\value{
Phase shifts in form of a vector that is utilized by the
\code{LSAPeriodogram}.
}
\description{
LSAPhase
}
