% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{M.Norm}
\alias{M.Norm}
\title{Math.NormalizeVector}
\usage{
M.Norm(vec)
}
\arguments{
\item{vec}{Input vector (or similar structure, e.g. N x 1 matrix).}
}
\value{
A unit vector parallel to the source one.
}
\description{
\code{M.Norm} normalizes vector, reducing its length in terms of Euclidian norm to 1.
}
\details{
Divides all source vector elements by a positive quantity selected in such a way that Euclidian norm of 
the resulting vector is 1. Requires a numerical structure coercible to 1D array.
}
