% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{map_lo}
\alias{map_lo}
\alias{imap_lo}
\alias{map2_lo}
\alias{pmap_lo}
\title{List_of mappers}
\usage{
map_lo(.x, .f, ...)

imap_lo(.x, .f, ...)

map2_lo(.x, .y, .f, ...)

pmap_lo(.l, .f, ...)
}
\arguments{
\item{.x, .y, .f, .l}{Parameters passed to respective \code{purrr} method}

\item{...}{Additional parameters}
}
\value{
A \code{vctrs::list_of} instead of a regular list.
}
\description{
List_of mappers
}
