% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by_ex.R
\name{order_by_ex}
\alias{order_by_ex}
\title{order_by_ex}
\usage{
order_by_ex(.x, .f = identity, ...)
}
\arguments{
\item{.x}{The object to sort}

\item{.f}{Accessor function that is evaluated for each element of \code{.x}}

\item{...}{\code{purrr}-style arguments to the function}
}
\value{
Ordered source collection
}
\description{
Orders a sortable object based on the accessor function
}
