% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_tools.R
\name{passes}
\alias{passes}
\title{passes}
\usage{
passes(x)
}
\arguments{
\item{x}{Test expression in the form of \code{is_tibble(mtcars)}}
}
\value{
Returns the result of the test.
}
\description{
Should be used in conjunction with \code{assertthat::assert_that};
provides a more detailed error explanation.
}
\examples{
\dontrun{assert_that(passes(is_tibble(mtcars)))}
}
