% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{pforeach}
\alias{pforeach}
\title{pforeach}
\usage{
pforeach(.data, ...)
}
\arguments{
\item{.data}{An implicit argument passed by a pipe operatpr \code{\%>\%}.}

\item{...}{Additional parameters passed to \code{foreach} as is.}
}
\description{
pforeach
}
