% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_cluster}
\alias{plan_cluster}
\title{plan_cluster}
\usage{
plan_cluster(
  ...,
  .default_seq = ~tweak(sequential),
  .default_multi = ~tweak(cluster, workers = .)
)
}
\arguments{
\item{...}{Number of workers per each dimension}

\item{.default_seq}{Default tweak method for sequential execution}

\item{.default_multi}{Default tweak method for parallel execution}
}
\value{
Nothing
}
\description{
plan_cluster
}
