% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Converters.R
\name{Ra2Degrees}
\alias{Ra2Degrees}
\title{Ra2Degrees}
\usage{
Ra2Degrees(x, sep = " ", simple = TRUE)
}
\arguments{
\item{x}{Vector of values.}

\item{sep}{Separator between hours, minutes and seconds.}

\item{simple}{If \code{TRUE}, return vector of converted values.
If \code{FALSE}, returns a \code{tibble} with more detatiled infomration.}
}
\value{
Either a vector of length \code{n} o a tibble \code{n x 5}.
}
\description{
Converts \code{R}ight \code{A}scention to degrees.
}
