% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_smart}
\alias{read_smart}
\title{read_smart}
\usage{
read_smart(path, ...)
}
\arguments{
\item{path}{Path to read from.}

\item{...}{Additional paramteres passed to either of
\code{feather::read_feather}, \code{readr::read_rds} or 
\code{readr::read_table2}.}
}
\value{
Nothing
}
\description{
Reads output in one of the avilable formats
based on the file extension.
}
