% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{ReadList}
\alias{ReadList}
\title{ReadList}
\usage{
ReadList(file)
}
\arguments{
\item{file}{Path to a file}
}
\value{
A list of indexes generated from file
}
\description{
Reads specifically formatted sequences of integers from a file.
}
