% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{RoundIntervalTo}
\alias{RoundIntervalTo}
\title{RoundIntervalTo}
\usage{
RoundIntervalTo(x, rnd)
}
\arguments{
\item{x}{An input oredered vector of size 2 (Interval lims).}

\item{rnd}{Rounding base.}
}
\value{
A vector of size 2, both limits of which are powers of \code{rnd}
}
\description{
RoundIntervalTo
}
