% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{sec_len}
\alias{sec_len}
\title{sec_len}
\usage{
sec_len(x)
}
\arguments{
\item{x}{Object of some length (a list, a vector, etc)}
}
\value{
Integer sequence from 1 to \code{length(x)}
}
\description{
sec_len
}
