% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGplot2Tools.R
\name{Segments2Points}
\alias{Segments2Points}
\title{Segments2Points}
\usage{
Segments2Points(.dt, x, y, xend, yend)
}
\arguments{
\item{.dt}{Input table.}

\item{x}{\code{x} column name.}

\item{y}{\code{y} column name.}

\item{xend}{\code{x} end column name.}

\item{yend}{\code{y} end column name.}
}
\value{
Trnasformed table.
}
\description{
Segments2Points
}
