% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_Segments2PointsEx.R
\name{Segments2PointsEx}
\alias{Segments2PointsEx}
\title{Segments2Points}
\usage{
Segments2PointsEx(.dt, ..., suffix = "2")
}
\arguments{
\item{.dt}{Input table.}

\item{...}{Column names to transform.}

\item{suffix}{Suffix of paired column names.}
}
\value{
Trnasformed table.
}
\description{
Segments2Points
}
