% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{seq_int_len}
\alias{seq_int_len}
\title{seq_int_len}
\usage{
seq_int_len(length.out)
}
\arguments{
\item{length.out}{Size of the sequence.}
}
\value{
Sequence of ints.
}
\description{
seq_int_len
}
