% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{SetMargins}
\alias{SetMargins}
\title{SetMargins}
\usage{
SetMargins(grob, type, margins)
}
\arguments{
\item{grob}{Grob to process.}

\item{type}{Type of the margin.
Allowed values: \code{"inner"} and \code{"outer"}.}

\item{margins}{Values of margins.}
}
\value{
Modified grob. Can be piped e.g. with \code{dplyr::`\%>\%`}.
}
\description{
SetMargins
}
