% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SourceAll.R
\name{SourceAll}
\alias{SourceAll}
\title{Sources files in the directory.
Useful for the initial project setup.}
\usage{
SourceAll(path, except, quiet = FALSE, recursive = TRUE)
}
\arguments{
\item{path}{Path where to look for .R files.}

\item{except}{Regex pattern to ignore files.}

\item{quiet}{If \code{FALSE}, prints message per each file.}

\item{recursive}{If \code{TRUE}, goes recurisvely into sub directories.}
}
\value{
Nothing
}
\description{
Sources files in the directory.
Useful for the initial project setup.
}
