% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTools.R
\name{Tex2Pdf}
\alias{Tex2Pdf}
\title{Tex2Pdf}
\usage{
Tex2Pdf(..., verbose = FALSE, additionalParams = "")
}
\arguments{
\item{...}{Source files.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints detailed output.}

\item{additionalParams}{Passed to 'pdflatex'.}
}
\description{
Tex2Pdf
}
