% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_2_pdf.R
\name{tex_2_pdf}
\alias{tex_2_pdf}
\alias{Tex2Pdf}
\title{Tex2Pdf}
\usage{
tex_2_pdf(..., verbose = FALSE, additionalParams = "")

Tex2Pdf(..., verbose = FALSE, additionalParams = "")
}
\arguments{
\item{...}{Source files.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints detailed output.}

\item{additionalParams}{Passed to 'pdflatex'.}
}
\description{
Tex2Pdf
}
