% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{TrimRibbonData}
\alias{TrimRibbonData}
\title{TrimRibbonData}
\usage{
TrimRibbonData(.data, x, y, xlim, ylim, lwr, upp, ...)
}
\arguments{
\item{.data}{Input table.}

\item{x}{\code{x} column.}

\item{y}{\code{y} column.}

\item{xlim}{\code{x} limits.}

\item{ylim}{\code{y} limits.}

\item{lwr}{Lower \code{y} limit of ribbon.}

\item{upp}{Upper \code{y} limit of ribbon}

\item{...}{Additional columns to preserve.}
}
\value{
Table ready to be plotted.
}
\description{
TrimRibbonData
}
