% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{unique_f}
\alias{unique_f}
\title{unique_f}
\usage{
unique_f(x, eps = 1L)
}
\arguments{
\item{x}{Vector to test.}

\item{eps}{Floating-point comparison tolerance}
}
\value{
Vector of unique items tkaen from the source vector.
}
\description{
unique_f
}
