% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{UniqueTol}
\alias{UniqueTol}
\title{UniqueTol}
\usage{
UniqueTol(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{Vector to check.}

\item{tol}{Tolerance level for comparisons.}
}
\value{
Unique elements of the vector.
}
\description{
UniqueTol
}
