% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_methods.R
\name{UniqueWhichTol}
\alias{UniqueWhichTol}
\title{UniqueWhichTol}
\usage{
UniqueWhichTol(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{Vector to check.}

\item{tol}{Tolerance level for comparisons.}
}
\value{
Indices of unique elements within given tlerance.
}
\description{
UniqueWhichTol
}
