% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_fixes.R
\name{vec_ptype_abbr.fs_path}
\alias{vec_ptype_abbr.fs_path}
\alias{vec_ptype2.fs_path}
\alias{vec_ptype2.fs_path.glue}
\alias{vec_ptype2.fs_path.default}
\alias{vec_ptype2.fs_path.character}
\alias{vec_ptype2.character.fs_path}
\alias{vec_ptype2.fs_path.fs_path}
\alias{vec_cast.fs_path}
\alias{vec_cast.fs_path.default}
\alias{vec_cast.fs_path.glue}
\alias{vec_cast.fs_path.character}
\alias{vec_cast.character.fs_path}
\title{fs_path {vctrs} methods}
\usage{
vec_ptype_abbr.fs_path(x, ...)

vec_ptype2.fs_path(x, y, ...)

\method{vec_ptype2}{fs_path.glue}(x, y, ...)

\method{vec_ptype2}{fs_path.default}(x, y, ...)

\method{vec_ptype2}{fs_path.character}(x, y, ...)

\method{vec_ptype2}{character.fs_path}(x, y, ...)

\method{vec_ptype2}{fs_path.fs_path}(x, y, ...)

vec_cast.fs_path(x, to, ...)

\method{vec_cast}{fs_path.default}(x, to, ...)

\method{vec_cast}{fs_path.glue}(x, to, ...)

\method{vec_cast}{fs_path.character}(x, to, ...)

\method{vec_cast}{character.fs_path}(x, to, ...)
}
\arguments{
\item{x, y}{Vectors to check/cast.}

\item{...}{Extra parameters.}

\item{to}{Target type.}
}
\description{
fs_path {vctrs} methods
}
