% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_fixes.R
\name{vec_ptype_abbr.glue}
\alias{vec_ptype_abbr.glue}
\alias{vec_ptype2.glue}
\alias{vec_ptype2.glue.default}
\alias{vec_ptype2.glue.fs_path}
\alias{vec_ptype2.glue.character}
\alias{vec_ptype2.character.glue}
\alias{vec_ptype2.glue.glue}
\alias{vec_cast.glue}
\alias{vec_cast.glue.default}
\alias{vec_cast.glue.fs_path}
\alias{vec_cast.glue.character}
\alias{vec_cast.character.glue}
\title{glue {vctrs} methods}
\usage{
vec_ptype_abbr.glue(x, ...)

vec_ptype2.glue(x, y, ...)

\method{vec_ptype2.glue}{default}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.glue}{fs_path}(x, y, ..., x_arg = "x", y_arg = "y")

\method{vec_ptype2.glue}{character}(x, y, ...)

\method{vec_ptype2.character}{glue}(x, y, ...)

\method{vec_ptype2.glue}{glue}(x, y, ...)

vec_cast.glue(x, to, ...)

\method{vec_cast.glue}{default}(x, to, ...)

\method{vec_cast.glue}{fs_path}(x, to, ...)

\method{vec_cast.glue}{character}(x, to, ...)

\method{vec_cast.character}{glue}(x, to, ...)
}
\arguments{
\item{x, y}{Vectors to check/cast.}

\item{...}{Extra parameters.}

\item{x_arg, y_arg}{Names of the parameters to include in the error message.}

\item{to}{Target type.}
}
\description{
glue {vctrs} methods
}
