% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{vec_assert_numeric}
\alias{vec_assert_numeric}
\title{vec_assert_numeric}
\usage{
vec_assert_numeric(x, size = NULL,
  arg = rlang::as_label(substitute(x)))
}
\arguments{
\item{x}{Vector to test}

\item{size}{Desired size. Can be \code{NULL}.}

\item{arg}{Arg names, defaults to the name of \code{X}.}
}
\value{
Invisibly returns \code{x} cast to \code{double()}
}
\description{
vec_assert_numeric
}
