% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{vec_cast_common_flatten}
\alias{vec_cast_common_flatten}
\title{vec_cast_common_flatten}
\usage{
vec_cast_common_flatten(x, y, .to = NULL)
}
\arguments{
\item{x}{Frist argument.}

\item{y}{Second argument.}

\item{.to}{Optional target type. If \code{NULL}, common type is inferred.}
}
\value{
List with two elements (\code{x} and \code{y}) that are source collections
cast to the common type and flattened.
}
\description{
vec_cast_common_flatten
}
