% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{vec_cast_integerish}
\alias{vec_cast_integerish}
\title{vec_cast_integerish}
\usage{
vec_cast_integerish(x, arg = rlang::as_label(x))
}
\arguments{
\item{x}{Vector to test.}

\item{arg}{Arg name for error message.}
}
\value{
\code{x} cast to \code{integer()} if \code{x} is coercible to \code{integer()}
i.e. \code{5.0} is converted to \code{5L}, while \code{5.1} triggers an exception.
}
\description{
vec_cast_integerish
}
