% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{vec_rbind_uq}
\alias{vec_rbind_uq}
\title{vec_rbind_uq}
\usage{
vec_rbind_uq(
  x,
  .ptype = NULL,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)
}
\arguments{
\item{x}{Input list, which is unquoted (!!!).}

\item{.ptype}{Type to cast to (if \code{NULL}, common type).}

\item{.names_to}{Where to put names.}

\item{.name_repair}{Name repair.}
}
\value{
Bind data object
}
\description{
vec_rbind_uq
}
