% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{\%within\%}
\alias{\%within\%}
\alias{\%withini\%}
\title{Within}
\usage{
lhs \%within\% rhs

lhs \%withini\% rhs
}
\arguments{
\item{lhs}{Vector to test}

\item{rhs}{Vector of size 2.}
}
\value{
Logical vector of length of \code{lhs}.
}
\description{
Within
}
