% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{WithinL}
\alias{WithinL}
\title{WithinL}
\usage{
WithinL(x, low, upp)
}
\arguments{
\item{x}{Input collection.}

\item{low}{Lower boundary.}

\item{upp}{Upper boundary.}
}
\value{
Logical vector.
}
\description{
Returns \code{TRUE}/\code{FALSE} vector indicating
which elements are within the range.
}
