% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_fixed}
\alias{write_fixed}
\alias{WriteFixed}
\title{write_fixed}
\usage{
write_fixed(frame, path, frmt, append = FALSE)

WriteFixed(frame, path, frmt, append = FALSE)
}
\arguments{
\item{frame}{\link{data.frame} or \link{tibble} to print.}

\item{path}{Path to the output file.}

\item{frmt}{\link{sprintf}-compatible format.
Either one value applied to all columns,
or a \code{character} vector of \code{ncol(frame)} elements.}

\item{append}{If \code{TRUE}, appends data to the existing file.}
}
\description{
Prints table to the target file in a fixed-format manner.
}
