% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_smart}
\alias{write_smart}
\title{write_smart}
\usage{
write_smart(data, path, ...)
}
\arguments{
\item{data}{Input table.}

\item{path}{Path to save.}

\item{...}{Additional paramteres passed to either of
\code{feather::write_feather}, \code{readr::write_rds} or 
\code{RLibs::WriteFixed}.}
}
\value{
Nothing
}
\description{
Writes output in one of the avilable formats
based on the file extension.
}
