% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteFixed}
\alias{WriteFixed}
\title{WriteFixed}
\usage{
WriteFixed(frame, path, frmt, append = FALSE)
}
\arguments{
\item{frame}{\link{data.frame} or \link{tibble} to print.}

\item{path}{Path to the output file.}

\item{frmt}{\link{sprintf}-compatible format.
Either one value applied to all columns,
or a \code{character} vector of \code{ncol(frame)} elements.}

\item{append}{If \code{TRUE}, appends data to the existing file.}
}
\description{
Prints table to the target file in a fixed-format manner.
}
