% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equal_f.R
\name{are_equal_f}
\alias{are_equal_f}
\title{are_equal_f}
\usage{
are_equal_f(x, y = rep(0, length(x)), tol = .Machine$double.eps)
}
\arguments{
\item{x}{First floating point collection.}

\item{y}{Second collection. By default has all zeroes
to enable shortcut comparison to zero.}

\item{tol}{Optional tolerance, defaulted to machine precision.}
}
\value{
Logical vector.
}
\description{
Checks for floating-point equality of two collections.
}
