% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R, R/data_frames.R
\name{clamp}
\alias{clamp}
\alias{Clamp}
\alias{clamp.numeric}
\alias{clamp.data.frame}
\alias{clamp.tbl_df}
\title{clamp}
\usage{
clamp(...)

Clamp(...)

\method{clamp}{numeric}(...)

\method{clamp}{data.frame}(...)

\method{clamp}{tbl_df}(...)
}
\arguments{
\item{...}{Parameter.}
}
\value{
Clamped numerics.
}
\description{
clamp
}
