% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{ClipXY}
\alias{ClipXY}
\title{ClipXY}
\usage{
ClipXY(.dt, x, y, xlim, ylim, ...)
}
\arguments{
\item{.dt}{Input table.}

\item{x}{\code{x} column name.}

\item{y}{code{y} column name.}

\item{xlim}{Limits on \code{x}.}

\item{ylim}{Limits on \code{y}.}

\item{...}{Additional columns to preserve.}
}
\description{
ClipXY
}
