% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools_GenerateBreaks.R
\name{GenerateBreaks}
\alias{GenerateBreaks}
\title{GenerateBreaks}
\usage{
GenerateBreaks(range, largeStep, smallStep, ticks, op = `*`)
}
\arguments{
\item{range}{Range of breaks (typically a scale of an axis).}

\item{largeStep}{Step for large breaks (the ones with labels).}

\item{smallStep}{If specified, used as step for small breaks.}

\item{ticks}{If \code{smallStep} is not specified, then \code{ticks} are used
to create small breaks.}

\item{op}{Used in combination with \code{ticks} as:
per each large break produces
\code{op(ticks, break)} small breaks and then limits them to \code{range}}
}
\value{
A list containing either $Large, $Small
collections of breaks, or both.
}
\description{
Generates large and small breaks within given range.
}
