% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{GenerateLog10Breaks}
\alias{GenerateLog10Breaks}
\title{GenerateLog10Breaks}
\usage{
GenerateLog10Breaks(ylim)
}
\arguments{
\item{ylim}{Axis limit.}
}
\value{
log10 breaks.
}
\description{
GenerateLog10Breaks
}
