% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlueFmt.R
\name{GlueFmt}
\alias{GlueFmt}
\alias{glue_fmt}
\title{GlueFmt}
\usage{
GlueFmt(..., .envir = parent.frame())

glue_fmt(..., .envir = parent.frame())
}
\arguments{
\item{...}{Passed to \code{glue::glue}.}

\item{.envir}{Evaluation environment.}
}
\value{
Format-aware interpoalted string.
}
\description{
GlueFmt
}
