% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{GrobMarginSet}
\alias{GrobMarginSet}
\title{GrobMarginSet}
\usage{
GrobMarginSet(grob, tLab, rLab, bLab, lLab, tAxis, rAxis, bAxis, lAxis)
}
\arguments{
\item{grob}{Input grob.}

\item{tLab}{Top label margin.}

\item{rLab}{Right label margin.}

\item{bLab}{Bottom label margin.}

\item{lLab}{Left label margin.}

\item{tAxis}{Top axis margin.}

\item{rAxis}{Right axis margin.}

\item{bAxis}{Bottom axis margin.}

\item{lAxis}{Left axis margin.}
}
\value{
Modified grob.
}
\description{
GrobMarginSet
}
