% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{GrobSetGaps}
\alias{GrobSetGaps}
\title{GrobSetGaps}
\usage{
GrobSetGaps(grob, top, right, bottom, left)
}
\arguments{
\item{grob}{Inout grob.}

\item{top}{Top gap.}

\item{right}{Right gap.}

\item{bottom}{Bottom gap.}

\item{left}{Left gap.}
}
\value{
Grob.
}
\description{
GrobSetGaps
}
