% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGPlot2Tools.R
\name{TrimLineData}
\alias{TrimLineData}
\title{TrimLineData}
\usage{
TrimLineData(.data, .var_x, .var_y, lim)
}
\arguments{
\item{.data}{Input table.}

\item{.var_x}{x variable.}

\item{.var_y}{y variable.}

\item{lim}{Limits.}
}
\description{
TrimLineData
}
