% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceMethods.R
\name{Within}
\alias{Within}
\title{Within}
\usage{
Within(x, range)
}
\arguments{
\item{x}{Input subset.}

\item{range}{Range within which elements of x should be.}
}
\value{
Subset of source vector x.
}
\description{
Returns a subset of x such as x is
within [min(range), max(range)].
}
